feat(Locale): Added proper locale support.
Added proper locale support to the plugin by including a new 'locale' option in config.yml. The default locale setting is 'en_US', and for now, 'en_US' and 'el' are the only available options. If you try to use another locale without the appropriate file, the plugin will not work. Additionally, a custom Locale class has been added to improve support for different locales.

Changed the SkinOptions serial number to 1L. From now on, the serial number will only change when the file is modified to prevent compatibility issues.